third_party/WebKit/LayoutTests/http/tests/misc/webtiming-buffer-full-no-event.html

<html>
<head>
<link rel="help" href="https://dvcs.w3.org/hg/webperf/raw-file/tip/specs/ResourceTiming/Overview.html#extensions-performance-interface">
<script src="../resources/js-test-pre.js"></script>
<script>
description("This test checks that Performance inherits EventTarget and that addEventListener() works for resourcetimingbufferfull events.");
window.jsTestIsAsync = true;
var bufferFullCount = 0;
var parameter;

function onBufferFull(event) {
    parameter = event;
    shouldBe('parameter.__proto__', 'Event.prototype');
    bufferFullCount++;
}

shouldBe('Performance.prototype.__proto__', 'EventTarget.prototype');

performance.webkitSetResourceTimingBufferSize(2);
shouldBeNull('performance.onwebkitresourcetimingbufferfull');
performance.addEventListener('webkitresourcetimingbufferfull', onBufferFull);
shouldBeNull('performance.onwebkitresourcetimingbufferfull');
</script>
</head>
<body>
<script>
function test() {
    // Make sure the onBufferFull callback was called exactly 1 time.
    shouldBe('bufferFullCount', '1');
    performance.removeEventListener('webkitresourcetimingbufferfull', onBufferFull);
    finishJSTest();
}

window.onload = test;
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
